+++ changelog.md
@@ -1,5 +1,5 @@
 ### General
 
-- Added carpet rule `dispensedItemNoRandomVelocity`
+- Added carpet rule `accurateDispenser`
 - Support for Minecraft 26.1-snapshot-1
 - Requires Fabric Loader 0.18.3+